+++ contrib/rpctest/main.go
@@ -16,8 +16,8 @@ import (
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/ethclient"
-	"github.com/zeta-chain/protocol-contracts/pkg/systemcontract.sol"
-	"github.com/zeta-chain/protocol-contracts/pkg/zetaeth.sol"
+	"github.com/zeta-chain/protocol-contracts-evm/pkg/systemcontract.sol"
+	"github.com/zeta-chain/protocol-contracts-evm/pkg/zetaeth.sol"
 )
 
 var (