+++ changelog.md
@@ -62,6 +62,7 @@ Also EnableSolanaAddressLookupTable feature flag should be set.
 ### Refactor
 
 * [4296](https://github.com/zeta-chain/node/pull/4296) - add zrepo package to zetaclient
+* [4356](https://github.com/zeta-chain/node/pull/4356) - rename protocol contract imports to `protocol-contracts-evm`
 
 ### Fixes